// fm-showcase/src/challenges/interactive-comments/types.ts
export interface User {
  username: string;
}

export interface Reply {
  id: number;
  content: string;
  createdAt: string;
  score: number;
  replyingTo: string;
  user: User;
}

export interface Comment {
  id: number;
  content: string;
  createdAt: string;
  score: number;
  user: User;
  replies: Reply[];
}

export interface CommentData {
  currentUser: User;
  comments: Comment[];
}